+++ app/src/release/kotlin/me/xizzhu/android/joshua/App.kt
@@ -17,10 +17,10 @@
 package me.xizzhu.android.joshua
 
 import android.app.Application
-import me.xizzhu.android.joshua.core.logger.Log
 import me.xizzhu.android.joshua.core.logger.android.CrashlyticsLogger
 import me.xizzhu.android.joshua.core.analytics.Analytics
 import me.xizzhu.android.joshua.core.analytics.android.FirebaseAnalyticsProvider
+import me.xizzhu.android.logger.Log
 
 abstract class BaseApp : Application() {
     override fun onCreate() {