+++ CHANGELOG.md
@@ -4,6 +4,8 @@ CHANGELOG
 #### Next Release
 - Fixed search result count
 - Migrated to use Logger library
+- Updated dependencies:
+  - Firebase Analytics and Core to 17.0.0
 
 #### v0.5.1 (2019-06-10)
 - Improved color for selected verses and verse details in night mode